FINISHED!

+++ src/Model/ExchangeRate.java
@@ -4,9 +4,9 @@ public final class ExchangeRate {
 
     private final Currency currencyFrom;
     private final Currency currencyTo;
-    private final float rate;
+    private final double rate;
 
-    public ExchangeRate(Currency currencyFrom, Currency currencyTo, float rate) {
+    public ExchangeRate(Currency currencyFrom, Currency currencyTo, double rate) {
         this.currencyFrom = currencyFrom;
         this.currencyTo = currencyTo;
         this.rate = rate;
@@ -20,7 +20,7 @@ public final class ExchangeRate {
         return currencyTo;
     }
 
-    public float getRate() {
+    public double getRate() {
         return rate;
     }
 }
